Separate interface and implementation
